# app/static/swagger.json
{
  "swagger": "2.0",
  "info": {
    "title": "SGHSS - API PARA GESTÃO HOSPITALAR",
    "description": "API RESTful para gestão de consultas médicas — médicos, pacientes e secretarías.",
    "version": "1.0.0"
  },
  "host": "127.0.0.1:5000",
  "basePath": "/api",
  "schemes": ["http"],
  "produces": ["application/json"],
  "consumes": ["application/json"],
  "securityDefinitions": {
    "BearerAuth": {
      "type": "apiKey",
      "name": "Authorization",
      "in": "header",
      "description": "Autenticação JWT. Use o formato: **Bearer {seu_token_aqui}**"
    }
  },

  "security": [
    {
      "BearerAuth": []
    }
  ],

  "paths": {
    "/login": {
      "post": {
        "tags": ["Autenticação"],
        "summary": "Login de usuário (gera token JWT válido por 60 minutos)",
        "description": "Permite que médicos, secretários e pacientes façam login e recebam um token JWT com validade de 60 minutos.",
        "consumes": ["application/json"],
        "produces": ["application/json"],
        "parameters": [
          {
            "in": "body",
            "name": "body",
            "description": "Credenciais de login do usuário.",
            "required": true,
            "schema": {
              "type": "object",
              "properties": {
                "NomeSobrenome": {
                  "type": "string",
                  "example": "DeleonSantos"
                },
                "Senha": {
                  "type": "string",
                  "example": "4556949"
                }
              },
              "required": ["nome", "senha"]
            }
          }
        ],
        "responses": {
          "200": {
            "description": "Login realizado com sucesso",
            "schema": {
              "type": "object",
              "properties": {
                "mensagem": { "type": "string", "example": "Login realizado com sucesso" },
                "token": { "type": "string", "example": "eyJ0eXAiOiJKV1QiLCJh..." },
                "expira_em": { "type": "string", "example": "60 minutos" }
              }
            }
          },
          "400": {
            "description": "Campos obrigatórios ausentes",
            "schema": {
              "type": "object",
              "properties": {
                "erro": { "type": "string", "example": "Email e senha são obrigatórios" }
              }
            }
          },
          "401": {
            "description": "Credenciais inválidas",
            "schema": {
              "type": "object",
              "properties": {
                "erro": { "type": "string", "example": "Credenciais inválidas" }
              }
            }
          }
        }
      }
    },
    "security": [
    {
      "BearerAuth": []
    }
  ],
  "/consulta/agenda_medica": {
      "get": {
        "tags": ["Medico"],
        "summary": "Consultar agenda do médico",
        "description": "Retorna todas as consultas do médico especificado, podendo filtrar por data.",
        "parameters": [
          
        ],
        "responses": {
          "200": { "description": "Lista de consultas encontradas" },
          "400": { "description": "Formato de data inválido" }
        }
      }
    },

    
    "/consulta/{consulta_id}/atendimento": {
      "post": {
        "tags": ["Medico"],
        "summary": "Finalizar consulta e registrar diagnóstico",
        "description": "Permite ao médico finalizar uma consulta e registrar o diagnóstico do paciente.",
        "parameters": [
          {
            "name": "consulta_id",
            "in": "path",
            "required": true,
            "type": "integer",
            "description": "ID da consulta a ser finalizada"
          },
          {
            "in": "body",
            "name": "body",
            "required": true,
            "schema": {
              "type": "object",
              "properties": {
                "diagnostico": { "type": "string" }
              },
              "required": ["diagnostico"]
            }
          }
        ],
        "responses": {
          "200": { "description": "Consulta finalizada com sucesso e diagnóstico registrado" },
          "400": { "description": "Diagnóstico não informado" },
          "404": { "description": "Consulta não encontrada" }
        }
      }
    },

    "/consulta/{consulta_id}/prescreve": {
      "post": {
        "tags": ["Medico"],
        "summary": "Prescrever medicamento",
        "description": "Permite ao médico prescrever um novo medicamento durante uma consulta.",
        "parameters": [
          {
            "name": "consulta_id",
            "in": "path",
            "required": true,
            "type": "integer",
            "description": "ID da consulta relacionada à prescrição"
          },
          {
            "in": "body",
            "name": "body",
            "required": true,
            "schema": {
              "type": "object",
              "properties": {
                "medicacao": { "type": "string" },
                "dosagem": { "type": "string" },
                "orientacoes": { "type": "string" }
                
              },
              "required": ["medicacao", "dosagem", "orientacoes"]
            }
          }
        ],
        "responses": {
          "201": { "description": "Medicamento prescrito com sucesso" },
          "400": { "description": "Campos obrigatórios faltando" },
          "404": { "description": "Consulta não encontrada" }
        }
      }
    },

    "/consulta/{consulta_id}/solicita_exame": {
      "post": {
        "tags": ["Medico"],
        "summary": "Solicitar exame para o paciente",
        "description": "Permite ao médico solicitar um exame durante o atendimento de uma consulta.",
        "parameters": [
          {
            "name": "consulta_id",
            "in": "path",
            "required": true,
            "type": "integer",
            "description": "ID da consulta relacionada ao exame"
          },
          {
            "in": "body",
            "name": "body",
            "required": true,
            "schema": {
              "type": "object",
              "properties": {
                "exame": { "type": "string" }
                
              },
              "required": ["Exame"]
            }
          }
        ],
        "responses": {
          "201": { "description": "Exame solicitado com sucesso" },
          "400": { "description": "Campos obrigatórios faltando" },
          "404": { "description": "Consulta não encontrada" }
        }
      }
    },
    
    "/consulta/{consulta_id}/medico_cancelamento": {
      "post": {
        "tags": ["Medico"],
        "summary": "Cancelar uma consulta",
        "parameters": [
          {
            "name": "consulta_id",
            "in": "path",
            "required": true,
            "type": "integer"
          }
        ],
        "responses": {
          "200": { "description": "Consulta cancelada com sucesso" },
          "404": { "description": "Consulta não encontrada" }
        }
      }
    },
 
    "/consulta/lista_medico_credenciado": {
      "get": {
        "tags": ["Paciente"],
        "summary": "Listar todos os médicos credenciados",
        "description": "Retorna uma lista de todos os médicos cadastrados, incluindo id, nome e especialidade.",
        "responses": {
          "200": {
            "description": "Lista de médicos retornada com sucesso",
            "schema": {
              "type": "array",
              "items": {
                "type": "object",
                "properties": {
                  "id": { "type": "integer", "example": 1 },
                  "nome": { "type": "string", "example": "Dra. Maria Magalhães" },
                  "especialidade": { "type": "string", "example": "Pediatria" }
                }
              }
            }
          },
          "403": {
            "description": "Paciente não identificado",
            "schema": {
              "type": "object",
              "properties": {
                "erro": { "type": "string", "example": "Paciente não identificado." }
              }
            }
          }
        }
        
      }
    },
    "/consulta/agendamento_paciente": {
      "get": {
        "tags": ["Paciente"],
        "summary": "Ver todas as consultas agendadas de um paciente",
        "parameters": [
          
        ],
        "responses": {
          "200": { "description": "Lista de consultas do paciente" },
          "400": { "description": "Erro de formato de data" }
        }
      }
    },

    "/consulta/agenda_medica/{medico_id}": {
      "get": {
        "tags": ["Paciente"],
        "summary": "Listar agenda de um médico por especialidade",
        "parameters": [
          {
            "name": "medico_id",
            "in": "path",
            "required": true,
            "type": "integer",
            "description": "ID do médico"
          },
          {
            "name": "data",
            "in": "query",
            "type": "string",
            "format": "date",
            "description": "Filtrar por data (YYYY-MM-DD)"
          }
        ],
        "responses": {
          "200": { "description": "Lista de horários disponíveis" },
          "400": { "description": "Formato de data inválido" }
        }
      }
    },

    "/consulta/novo_agendamento": {
      "post": {
        "tags": ["Paciente"],
        "summary": "Agendar uma nova consulta",
        "parameters": [
          {
            "in": "body",
            "name": "body",
            "required": true,
            "schema": {
              "type": "object",
              "properties": {
                "paciente_id": { "type": "integer" },
                "medico_id": { "type": "integer" },
                "data": { "type": "string", "format": "date" },
                "hora": { "type": "string", "format": "time" }
              },
              "required": ["paciente_id", "medico_id", "data", "hora"]
            }
          }
        ],
        "responses": {
          "201": { "description": "Consulta agendada com sucesso" },
          "400": { "description": "Erro de validação" }
        }
      }
    },

    "/consulta/{consulta_id}/paciente_cancelamento": {
      "post": {
        "tags": ["Paciente"],
        "summary": "Cancelar uma consulta",
        "parameters": [
          {
            "name": "consulta_id",
            "in": "path",
            "required": true,
            "type": "integer"
          }
        ],
        "responses": {
          "200": { "description": "Consulta cancelada com sucesso" },
          "404": { "description": "Consulta não encontrada" }
        }
      }
    },
    
    
    
    "/cadastra/medico": {
      "post": {
        "tags": ["Secretaria"],
        "summary": "Cadastrar novo médico",
        "parameters": [
          {
            "in": "body",
            "name": "body",
            "required": true,
            "schema": {
              "type": "object",
              "properties": {
                "nome": { "type": "string" },
                "especialidade": { "type": "string" },
                "crm": { "type": "string" },
                "usuario": { "type": "string" },
                "senha": { "type": "string" },
                "nivel_acesso": { "type": "string" }

              },
              "required": ["nome", "especialidade", "crm", "usuario", "senha", "nivel_acesso"]
            }
          }
        ],
        "responses": {
          "201": { "description": "Médico cadastrado com sucesso" },
          "400": { "description": "Campos obrigatórios faltando" }
        }
      }
    },

    "/consulta/medico/{medico_id}": {
      "delete": {
        "tags": ["Secretaria"],
        "summary": "Excluir médico pelo ID",
        "parameters": [
          {
            "name": "medico_id",
            "in": "path",
            "required": true,
            "type": "integer"
          }
        ],
        "responses": {
          "200": { "description": "Médico excluído com sucesso" },
          "404": { "description": "Médico não encontrado" }
        }
      }
    },

    "/cadastra/paciente": {
      "post": {
        "tags": ["Secretaria"],
        "summary": "Cadastrar novo paciente",
        "parameters": [
          {
            "in": "body",
            "name": "body",
            "required": true,
            "schema": {
              "type": "object",
              "properties": {
                "nome": { "type": "string" },
                "cpf": { "type": "string" },
                "data_nascimento": { "type": "string", "format": "date" },
                "usuario": { "type": "string" },
                "senha": { "type": "string" },
                "nivel_acesso": { "type": "string" }
              },
              "required": ["nome", "cpf", "data_nascimento", "usuario", "senha", "nivel_acesso"]
            }
          }
        ],
        "responses": {
          "201": { "description": "Paciente cadastrado com sucesso" },
          "400": { "description": "Campos obrigatórios faltando" }
        }
      }
    },

    "/consulta/paciente/{paciente_id}": {
      "delete": {
        "tags": ["Secretaria"],
        "summary": "Excluir paciente pelo ID",
        "parameters": [
          {
            "name": "paciente_id",
            "in": "path",
            "required": true,
            "type": "integer"
          }
        ],
        "responses": {
          "200": { "description": "Paciente excluído com sucesso" },
          "404": { "description": "Paciente não encontrado" }
        }
      }
    },

    "/consulta/cadastro_usuarios": {
      "get": {
        "tags": ["Secretaria"],
        "summary": "Listar todos os usuários do sistema",
        "description": "Retorna todos os usuários cadastrados (pacientes, médicos e secretarias), com opção de filtrar por tipo.",
        "parameters": [
          {
            "name": "tipo",
            "in": "query",
            "type": "string",
            "enum": ["paciente", "medico", "secretaria"],
            "description": "Filtrar usuários por tipo"
          }
        ],
        "responses": {
          "200": { "description": "Lista de usuários retornada com sucesso" },
          "403": { "description": "Acesso não autorizado" }
        }
      }
    },
    "/consulta/consultas_marcadas": {
      "get": {
        "tags": ["Secretaria"],
        "summary": "Listar todas as consultas agendadas",
        "parameters": [
          {
            "name": "status",
            "in": "query",
            "type": "string",
            "enum": ["Agendada", "Realizada", "Cancelada"],
            "description": "Filtrar consultas por status"
          }
        ],
        "responses": {
          "200": { "description": "Lista de consultas retornada com sucesso" }
        }
      }
    }
  }
}